app/requirements.txt
FastAPI
uvicorn
gradio
python-dotenv
langchain-core
langchain-openai
pydantic
logging